Extract and fix batch transfer tests
